backend/requirements.txt
alembic==1.13.1
asgiref==3.7.2
asn1crypto==1.5.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distlib==0.3.8
Django==5.0.1
docopt==0.6.2
dominate==2.5.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.1.2
MarkupSafe==2.1.5
mysql-client==0.0.1
mysql-connector==2.2.9
mysqlclient==2.2.1
packaging==23.2
pillow==10.2.0
pipenv==2023.12.1
pipreqs==0.4.13
platformdirs==3.11.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==0.9.3
pyOpenSSL==23.3.0
python-dateutil==2.8.1
python-dotenv==1.0.1
python-editor==1.0.4
pytz==2024.1
requests==2.31.0
setuptools==69.0.3
six==1.14.0
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.51
sqlparse==0.4.4
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.25.0
visitor==0.1.3
Werkzeug==3.0.1
WTForms==2.2.1
yarg==0.1.9
